backend/requirements.txt
# Project Dwight - Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI/ML
openai==1.12.0
langchain==0.1.5
langchain-openai==0.0.5
langchain-community==0.0.17
tiktoken==0.5.2

# Vector Store
faiss-cpu==1.7.4

# Data Processing
numpy==1.26.3
pandas==2.2.0

# Environment & Config
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Google Sheets Integration
gspread==6.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# Email
aiosmtplib==3.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
